chore: add value prop table view and value prop button view
